Reorganise the default cipher preference order.
After all the side-channel removal work I've done, I'm demoting
Blowfish and Arcfour right to the bottom of the pile, because they're
the ciphers that still have cache and timing leaks in them. (Neither
one can be sanitised without making it _extremely_ slow - the factor
of 4.5 slowdown in DES would be small by comparison).

Single-DES is also still below the warning threshold for simply being
too weak (irrespective of side channels), but it's above the other two.